+++ docs/changes.rst
@@ -16,6 +16,7 @@ Unreleased.
 
  - Added :meth:`Image.affine() <wand.image.BaseImage.affine>` method.
  - Added :meth:`Image.bilateral_blur() <wand.image.BaseImage.bilateral_blur>` method.
+ - Added :meth:`Image.channel_fx() <wand.image.Image.channel_fx>` method.
  - Added :meth:`Image.evaluate_images() <wand.image.BaseImage.evaluate_images>` method.
  - Added :meth:`Image.floodfill() <wand.image.BaseImage.floodfill>` method.
  - Added :meth:`Image.morph() <wand.image.Image.morph>` method.